gcp_dev_vdi/requirements.txt
google-cloud-storage
google-cloud-compute
PyYAML
